add blockhash poller & transaction sender
